scripts/requirements.txt
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
jupyter==1.0.0
matplotlib==3.8.0
seaborn==0.13.0 